types to proper files
